rename package name tests on infludb publisher
